# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing veterinary clinic operations including appointments, medical records, treatments, and more.",
        "title": "Veterinary Management System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.veterinary-system.io/support",
            "email": "support@veterinary-system.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/veterinarians": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all veterinarians in the clinic",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Veterinarians"
                ],
                "summary": "Get all veterinarians",
                "responses": {
                    "200": {
                        "description": "List of veterinarians",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Veterinarian"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new veterinarian to the clinic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Veterinarians"
                ],
                "summary": "Create a new veterinarian",
                "parameters": [
                    {
                        "description": "Veterinarian details",
                        "name": "veterinarian",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Veterinarian"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veterinarian created successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.Veterinarian"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/veterinarians/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a specific veterinarian",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Veterinarians"
                ],
                "summary": "Get veterinarian by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Veterinarian ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veterinarian details",
                        "schema": {
                            "$ref": "#/definitions/domain.Veterinarian"
                        }
                    },
                    "400": {
                        "description": "Invalid veterinarian ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update information for a specific veterinarian",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Veterinarians"
                ],
                "summary": "Update veterinarian by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Veterinarian ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Veterinarian details",
                        "name": "veterinarian",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Veterinarian"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veterinarian updated successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.Veterinarian"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Veterinarian not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a veterinarian from the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Veterinarians"
                ],
                "summary": "Delete veterinarian by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Veterinarian ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veterinarian deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid veterinarian ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Veterinarian not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Veterinarian": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "specialty": {
                    "type": "string"
                },
                "veterinarian_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}